src/student_attendance_management_system/requirements.txt
asgiref==3.2.3
autopep8==1.5
Django==3.0.3
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
pycodestyle==2.5.0
pytz==2019.3
sqlparse==0.3.0
